added Readme.md, added some comments, improved file structure
