+++ relational/student_projects/2021_mannhardt/benchmarks/baselineBench.json
@@ -1,13 +1,13 @@
 {
     "type": "coverage",
-    "data.test": "Dataset.csv",
-    "data.pre_train": "Dataset.csv",
+    "data.test": "../Dataset.csv",
+    "data.pre_train": "../Dataset.csv",
     "corresponding_data": true,
     "domains": ["syllogistic", "spatial-relational"],
     "response_types": ["single-choice", "verify"],
     "models": [
-        "models/TCModel.py",
-        "models/MFA.py",
-        "models/simpleModels/random_uniform.py"
+        "../models/TCModel.py",
+        "../models/MFA.py",
+        "../models/simpleModels/random_uniform.py"
     ]
 }
