--- relational/student_projects/2021_mannhardt/machineLearningBench.json
@@ -1,16 +1,0 @@
-{
-    "type": "coverage",
-    "data.test": "Dataset.csv",
-    "data.pre_train": "Dataset.csv",
-    "corresponding_data": true,
-    "domains": ["syllogistic", "spatial-relational"],
-    "response_types": ["single-choice", "verify"],
-    "models": [
-        "models/TCModel.py",
-        "models/ensembleModel.py",
-        "models/MFA.py",
-        "models/IBCF.py",
-        "models/UBCF.py",
-        "models/mlp"
-    ]
-}
