+++ relational/student_projects/2021_mannhardt/models/MFA.py
@@ -3,6 +3,7 @@ import numpy as np
 import ccobra
 from taskencoder import *
 
+
 class MFA(ccobra.CCobraModel):
     def __init__(self, name='MFA'):
         super(MFA, self).__init__(name, ["spatial-relational"], ["single-choice", "verify"])
