+++ relational/student_projects/2021_mannhardt/models/taskencoder.py
@@ -1,61 +1,10 @@
-informations = ["id", "cat", "cons", "blv", "prems", "cormodel", "cfact", "revplausible", "revlo"]
 
-class TaskEncoder():
-    def __init__(self, task):
-        self.task = task
+'''
+inclues helper functions used by some models
+'''
 
-        # initial premises
-        self.prem1 = ""
-        self.prem2 = ""
-
-        # initial models
-        self.modelLeft = ""
-        self.modelRight = ""
-
-        # counterfact
-        self.counterfact = ""
-
-        # model revision choices
-        self.modelChoiceLeft = ""
-        self.modelChoiceLeft = ""
-    
-    def encode_task(self):
-        """Returns dictionary with all task information given in task string
-        e.g. '45_N_Incon_ug_aufauf_modri_factab_revgle_lori_blank' turns to
-        """
-        taskList = self.task.split('_')
-        taskDict = dict()
-        for i in range(len(informations)):
-            inf = informations[i]
-            taskInf = taskList[i]
-            if taskInf.endswith("le"):
-                taskDict[inf] = "left"
-                continue
-            elif taskInf.endswith("ri"):
-                taskDict[inf] = "right"
-                continue
-            else: 
-                taskDict[inf] = taskInf
-        return taskDict
-
-def encode_task(task):
-    """Returns dictionary with all task information given in task string
-    e.g. '45_N_Incon_ug_aufauf_modri_factab_revgle_lori_blank' turns to
-    """
-    taskList = task.split('_')
-    taskDict = dict()
-    for i in range(len(informations)):
-        inf = informations[i]
-        taskInf = taskList[i]
-        if taskInf.endswith("le"):
-            taskDict[inf] = "left"
-            continue
-        elif taskInf.endswith("ri"):
-            taskDict[inf] = "right"
-            continue
-        else: 
-            taskDict[inf] = taskInf
-    return taskDict
+import numpy as np
+import math
 
 def list_to_string_help(stringList):
     for i in stringList:
@@ -64,8 +13,51 @@ def list_to_string_help(stringList):
         elif isinstance(i, list):
             yield list_to_string(i)
 
+
 def list_to_string(stringList):
     fullStr = ""
     for s in list_to_string_help(stringList):
         fullStr += s
     return fullStr
+
+
+def keywithmaxval(d):
+     """ return the key with the max value
+         src: stackoverflow
+         """  
+     v=list(d.values())
+     k=list(d.keys())
+     return k[v.index(max(v))]
+
+def task_to_string(syntaxTree):
+    """
+    Changes e.g. ["Left;A;B", "Right";"C";"B"] to "LeftABRightCB"
+    Args:
+        task
+    Returns:
+        string - task written as string
+    """
+    if isinstance(syntaxTree[0][0], bool):
+        return syntaxTree[0][0]
+
+    if isinstance(syntaxTree, str) or isinstance(syntaxTree, bool):
+        return syntaxTree
+
+    all = ""
+    for elem in syntaxTree:
+        all += task_to_string(elem)
+    return all
+
+
+def dotproduct(v1, v2):
+    return sum((a*b) for a, b in zip(v1, v2))
+
+
+def length(v):
+    return math.sqrt(dotproduct(v, v))
+
+
+def angle(v1, v2):
+    normalized_v1 = v1/np.linalg.norm(v1)
+    normalized_v2 = v2/np.linalg.norm(v2)
+    return round(math.acos(dotproduct(normalized_v1, normalized_v2) / (length(normalized_v1) * length(normalized_v2))), 5)
